% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trex.R
\name{Trex}
\alias{Trex}
\title{Transformation of relative expression}
\usage{
Trex(x, y, ctrl = "control", usemc = FALSE)
}
\arguments{
\item{x}{A vector with the trait values (observations) that should be transformed. Should have the same length as treat.}

\item{y}{A vector containing the treatment for each observation. Should have the same length as trait.}

\item{ctrl}{The name of the factor for your control group as to be found in treat. This is important for standardization.}

\item{usemc}{Can be either FALSE (default) or TRUE. In case of FALSE, the values are standardized by the span of
treatment and control (max-min)). In case of TRUE., mean control is used to standardize the values. In this case,
values will be standardized for the experiment. Therefore, values are completely independent of the experimental
design and other factors. This increases the comparability with other studies, regarding the original values.
However, when using this for calculation of the Index for Adaptive Responses, the weighing (dmax) of the traits
will be TRUE. in each case. Therefore, no real weighing takes place. This might be desired, if the InARes framework
is used as plasticity index.}
}
\value{
Returns a vector containing the transformed data.
}
\description{
Part of the Index for Adaptive Responses (InARes).
Transforms trait values of a treatment (or similar data) in relation to a control (or similar data).
This results in transformed values that reflect the relative expression (therefore rex) of a trait of
a treated organism in relation to a control organism. They are standardized and normalized.
The transformation is calculated as:
(value - mean of control values) / (max - min of the control and value treatments)
This transformation is the first step of calculating the Index for Adaptive Responses, followed by
the functions InARes() and (optionally) dContribution().
}
\examples{
# load example data
library(InARes)
mydata <- data.InARes

# transform one trait for calculating the Index for Adaptive Responses in a next step
mydata$BL_rex <- Trex(x = mydata$BL,
                      y = mydata$induced,
                      ctrl = "n")

# transform a set of traits for calculating the Index for Adaptive Responses in a next step
params <- c("BL", "SL", "BW", "BWL", "Forn", "Furca", "SBAd", "SBAv", "meansld", "meanslv")
for(i in 1:length(params)){
  mydata[, paste(params[i], "_rex", sep = "")] <- Trex(x = mydata[,params[i]],
                                                       y = mydata$induced,
                                                       ctrl = "n")
}

# transform one trait to use it in a plasticity index
mydata$BL_uw <- Trex(x = mydata$BL,
                     y = mydata$induced,
                     ctrl = "n",
                     usemc = 1)


}
\references{
TODO: Link to the Publication
}
\seealso{
InARes(), calculates a weighted mean of all rex transformed trait values.

dContribution(), can be used to estmate each traits' contribution to the Index of Adaptive Responses.
}
