% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InARes.R
\name{InARes}
\alias{InARes}
\title{Calculation of the Index for Adaptive Responses}
\usage{
InARes(
  data,
  traits.adapt,
  traits.mal = FALSE,
  treatCol = "treatment",
  ctrl = "control",
  as.defined = FALSE,
  dmax.adapt = FALSE,
  dmax.mal = FALSE,
  na.action = "keep",
  as.PI = FALSE
)
}
\arguments{
\item{data}{A data.frame to be used.}

\item{traits.adapt}{The names of the columns that contains the d transformed values
of the traits to be incorporated where an increase is considered an adaptation.}

\item{traits.mal}{The names of the columns that contains the d transformed values
of the traits to be incorporated where an increase is considered as a maladaptation.}

\item{treatCol}{column name of the treatments (defaults to 'treatment'). At the current state, everything that is not control is
considered as one treatment. If there are several treatments in the data set, we recommend to subset the data accordingly.}

\item{ctrl}{the name of the control treatment as set in treatCol (defaults to 'control')}

\item{as.defined}{if TRUE the sign of dmax will be positive for traits.adapt and negative
for traits.mal when used as weighing parameter. If set to FALSE (default) the sign might be
positive or negative for traits.adapt, but always negative for traits.mal.
This parameter is thought for cases where the functionality
of the traits are unclear and the Index will be evaluated using dContribution().}

\item{dmax.adapt}{a vector containing values for dmax for each trait provided in traits.adapt.
It can be specified, if one wants to calculate values comparable with other experiments or the literature.
dmax has to be set for all traits or an NA should be included. Order should be accordingly as set
in traits.adapt. Each value must be between -1 and 1. Defaults to FALSE.}

\item{dmax.mal}{a vector containing values for dmax for each trait provided in traits.mal
(works similar as for dmax.adapt). Order should be accordingly as set in traits.adapt.
Each value must be between -1 and 1. Defaults to FALSE.}

\item{na.action}{Defines the way, how NAs are treated (row wise). Can be either "omit", "keep", or "interpolate".
Defaults to "keep".
When "keep" the Defence-Index value will be NA, if one or more of the trait values contain NA. Usually
recommended for unbiased estimates!
In case of "interpolate", it will use the median value of the trait of the current treatment. However,
this might bias your estimates in some way, if the median of the treatment in this trait is not a´fitting adequately,
but it helps if a lot of NAs occur in fields of different traits. We recommend this way, as a conservative
way to prevent NAs. However, if there are a lot of gaps in your trait data, it may bias the outcome strongly.
When "omit" it calculates the Defence-Index with traits where an NA occurs. In this case, NA-traits are
ignored and number of traits reduced by the number of traits with NA (for each individual). This, however,
may bias the outcome or the later estimation of each traits' contribution. The values may be misleading.
This way is not recommended and should only be used if you are certain about what you do!}

\item{as.PI}{can be set TRUE if the aim is to use this function for calculating an overall plasticity index.
In this case, all traits should be transformed with dTransform(..., usemc = 1). Additionally, all traits
should be provided via traits.adapt (not traits.mal) and as.defined should be set TRUE. Now you will receive
values that are measures of the mean deviation over all traits from the control mean.}
}
\value{
Returns a list containing the transformed data and additional information like d(max).
}
\description{
Calculates the Index for Adaptive Responses from several rex transformed traits:
if as.defined = TRUE: D = sum((-1) * d(x) / max( |d|(all treatments) )) / number of parameter
or
if as.defined = FALSE (default): D = sum((-1) * d(x) / ( max( |d|(all treatments) )) * median(treatment) / |median(treatment)| ) / number of parameter
The (-1) is only necessary, if we believe that the increase of one trait is maladaptive.
dmax will be 1 when usemc = 1 was selected in d transformation.
At the current state, everything that is not control is considered as one treatment. If there are several
treatments in the data set, we recommend to subset the data accordingly.
}
